create types file and add typecheck scripts
