+++ package.json
@@ -13,6 +13,9 @@
     "test": "jest",
     "lint": "eslint ./src",
     "lint:fix": "eslint ./src --fix",
+    "typecheck:esm": "tsc -p ./tsconfig.esm.json --noEmit",
+    "typecheck:cjs": "tsc -p ./tsconfig.cjs.json --noEmit",
+    "typecheck": "npm run typecheck:esm && npm run typecheck:cjs",
     "format": "prettier . --write"
   },
   "exports": {
