+++ src/index.ts
@@ -3,3 +3,4 @@ export * from "./GatewaySession";
 export * from "./RequestProgress";
 export * from "./Responder";
 export * from "./Util";
+export * from "./types";
